Prevent the device from being identified as a CDC device
Sets bDeviceClass, bDeviceSubClass, bDeviceProtocol,
bInterfaceClass, bInterfaceSubClass and bInterfaceProtocol
to 0 and remove all functional descriptors.
This means the ODrive is no longer recognized as
communications device class (CDC) device, thus avoiding
problems where the OS or other applications try to
communicate with or hog the device.

Related resources:
http://www.microchip.com/forums/m789836.aspx
https://www-user.tu-chemnitz.de/~heha/viewchm.php/hs/usb.chm/usb5.htm

Tested on Linux: the device no longer shows up as tty.
To be tested on macOS and Windows.